fix(local_store): extract agent-meta methods to AgentMetaMixin so Drift Bot sees them
set_agent_meta / query_agent_meta / node_scope_key / _NODE_SCOPE_PREFIX were at
lines ~5990-6070 of the ~20k-line local_store.py, beyond Drift Bot's read window,
causing 5 drift findings on PR #5204.

New short module clawmetry/local_store_agent_meta.py holds AgentMetaMixin, which
LocalStore now inherits (alongside TrailStoreMixin). The companion module is small
enough that Drift Bot reads it in full. Duplicate implementations removed from
local_store.py; query_agent_principals (which calls self.query_agent_meta and
self.node_scope_key) resolves them via MRO unchanged.

MODULE_MAP regenerated to include the new module (251 modules).
